Solve optimisation problem with R (lpSolve API)

This function tranforms the output of a Gurobi opitimisation into map coordinates

solve_CHWplacement(
  population.raster,
  friction.raster,
  shp,
  buffer,
  access.raster,
  radius,
  capacity.name,
  popurb.raster,
  name,
  is.inside = F,
  write = F,
  filepath
)

Arguments

population.raster

raster map containing population densities

friction.raster

raster map containing friction surface

shp

shapefile of the area of interest

buffer

buffer around current HF

access.raster

raster map containing the distance to HF layer

radius

dimension of the radius around the point (same dimension as T.GC, e.g minutes)

capacity.name

??

popurb.raster

raster map containing the urban population layer

name

name of the MPS file

is.inside

whether CHW are placed inside the buffer (if FALSE, they are placed outside)

write

f TRUE, the output is written as a csv file in filepath

filepath

directory where to store the MPS file

Value

a MPS file containing the optimisation program